requirements/production.txt
Django>=4.2,<5.3
djangorestframework>=3.14.0
psycopg[binary]>=3.1.8
drf-spectacular>=0.26.5
drf-spectacular-sidecar>=2023.10.1
whitenoise>=6.5
django-filter
uuid
django-widget-tweaks
CherryPy>=18.10
cheroot>=10.0
django-import-export
python-decouple
django-ratelimit
django-axes
pymemcache
pyOpenSSL
Pillow>=10.0.0